Add files via upload
Project 1 - ColumbiaX - Artificial Intelligence Course through www.edx.org
Solving 8-puzzle game with three different search algorithms. 
I have to point out that this version meets the requirements of the project and is graded with the result of 98.5/100.
One minor mistake is somewhere when you calculate maximum_depth of DFS algorithm, should be a something like a special case because most of the time it calculates the maximum_depth as it should.